Fix bug in accuracy calculation
The result of `torch.sum(preds == targets)` could be as much as `self.BATCH_SIZE`, so `len(data_loader)` isn't enough of a denominator.

There could be batches from the `data_loader` with less than `BATCH_SIZE`, so a running counter of would be more accurate, but I think this is good enough.

Signed-off-by: John Savage <jsavage@overstock.com>